+++ staging/vhost-device-sound/src/audio_backends/alsa.rs
@@ -13,7 +13,7 @@ use std::{
 
 use alsa::{
     pcm::{Access, Format, HwParams, State, PCM},
-    Direction, PollDescriptors, ValueOr,
+    PollDescriptors, ValueOr,
 };
 use virtio_queue::Descriptor;
 use vm_memory::Bytes;
@@ -22,18 +22,25 @@ use super::AudioBackend;
 use crate::{
     device::ControlMessage,
     stream::{PCMState, Stream},
-    virtio_sound::{
-        self, VirtioSndPcmSetParams, VIRTIO_SND_D_INPUT, VIRTIO_SND_D_OUTPUT, VIRTIO_SND_S_BAD_MSG,
-        VIRTIO_SND_S_NOT_SUPP,
-    },
-    Result as CrateResult,
+    virtio_sound::{self, VirtioSndPcmSetParams, VIRTIO_SND_S_BAD_MSG, VIRTIO_SND_S_NOT_SUPP},
+    Direction, Result as CrateResult,
 };
+
+impl From<Direction> for alsa::Direction {
+    fn from(val: Direction) -> Self {
+        match val {
+            Direction::Output => Self::Playback,
+            Direction::Input => Self::Capture,
+        }
+    }
+}
 
 type AResult<T> = std::result::Result<T, alsa::Error>;
 
 #[derive(Clone, Debug)]
 pub struct AlsaBackend {
     sender: Arc<Mutex<Sender<AlsaAction>>>,
+    streams: Arc<RwLock<Vec<Stream>>>,
 }
 
 #[derive(Debug)]
@@ -42,9 +49,7 @@ enum AlsaAction {
     Prepare(usize),
     Release(usize, ControlMessage),
     Start(usize),
-    Stop(usize),
-    Write(usize),
-    Read(usize),
+    DoWork(usize),
 }
 
 fn update_pcm(
@@ -55,39 +60,30 @@ fn update_pcm(
     *pcm_.lock().unwrap() = {
         let streams = streams.read().unwrap();
         let s = &streams[stream_id];
-        let pcm = PCM::new(
-            "default",
-            match s.direction {
-                d if d == VIRTIO_SND_D_OUTPUT => Direction::Playback,
-                d if d == VIRTIO_SND_D_INPUT => Direction::Capture,
-                other => panic!("Invalid virtio-sound stream: {}", other),
-            },
-            false,
-        )?;
+        let pcm = PCM::new("default", s.direction.into(), false)?;
 
         {
+            let rate = match s.params.rate {
+                virtio_sound::VIRTIO_SND_PCM_RATE_5512 => 5512,
+                virtio_sound::VIRTIO_SND_PCM_RATE_8000 => 8000,
+                virtio_sound::VIRTIO_SND_PCM_RATE_11025 => 11025,
+                virtio_sound::VIRTIO_SND_PCM_RATE_16000 => 16000,
+                virtio_sound::VIRTIO_SND_PCM_RATE_22050 => 22050,
+                virtio_sound::VIRTIO_SND_PCM_RATE_32000 => 32000,
+                virtio_sound::VIRTIO_SND_PCM_RATE_44100 => 44100,
+                virtio_sound::VIRTIO_SND_PCM_RATE_48000 => 48000,
+                virtio_sound::VIRTIO_SND_PCM_RATE_64000 => 64000,
+                virtio_sound::VIRTIO_SND_PCM_RATE_88200 => 88200,
+                virtio_sound::VIRTIO_SND_PCM_RATE_96000 => 96000,
+                virtio_sound::VIRTIO_SND_PCM_RATE_176400 => 176400,
+                virtio_sound::VIRTIO_SND_PCM_RATE_192000 => 192000,
+                virtio_sound::VIRTIO_SND_PCM_RATE_384000 => 384000,
+                _ => unreachable!(),
+            };
             let hwp = HwParams::any(&pcm)?;
             hwp.set_channels(s.params.channels.into())?;
-            hwp.set_rate(
-                match s.params.rate {
-                    virtio_sound::VIRTIO_SND_PCM_RATE_5512 => 5512,
-                    virtio_sound::VIRTIO_SND_PCM_RATE_8000 => 8000,
-                    virtio_sound::VIRTIO_SND_PCM_RATE_11025 => 11025,
-                    virtio_sound::VIRTIO_SND_PCM_RATE_16000 => 16000,
-                    virtio_sound::VIRTIO_SND_PCM_RATE_22050 => 22050,
-                    virtio_sound::VIRTIO_SND_PCM_RATE_32000 => 32000,
-                    virtio_sound::VIRTIO_SND_PCM_RATE_44100 => 44100,
-                    virtio_sound::VIRTIO_SND_PCM_RATE_48000 => 48000,
-                    virtio_sound::VIRTIO_SND_PCM_RATE_64000 => 64000,
-                    virtio_sound::VIRTIO_SND_PCM_RATE_88200 => 88200,
-                    virtio_sound::VIRTIO_SND_PCM_RATE_96000 => 96000,
-                    virtio_sound::VIRTIO_SND_PCM_RATE_176400 => 176400,
-                    virtio_sound::VIRTIO_SND_PCM_RATE_192000 => 192000,
-                    virtio_sound::VIRTIO_SND_PCM_RATE_384000 => 384000,
-                    _ => 44100,
-                },
-                ValueOr::Nearest,
-            )?;
+            hwp.set_rate(rate, ValueOr::Nearest)?;
+            hwp.set_rate_resample(false)?;
             hwp.set_format(match s.params.format {
                 virtio_sound::VIRTIO_SND_PCM_FMT_IMA_ADPCM => Format::ImaAdPCM,
                 virtio_sound::VIRTIO_SND_PCM_FMT_MU_LAW => Format::MuLaw,
@@ -114,17 +110,63 @@ fn update_pcm(
                 virtio_sound::VIRTIO_SND_PCM_FMT_DSD_U16 => Format::DSDU16LE,
                 virtio_sound::VIRTIO_SND_PCM_FMT_DSD_U32 => Format::DSDU32LE,
                 virtio_sound::VIRTIO_SND_PCM_FMT_IEC958_SUBFRAME => Format::iec958_subframe(),
-                _ => Format::Unknown,
+                _ => unreachable!(),
             })?;
 
             hwp.set_access(Access::RWInterleaved)?;
 
-            // > A period is the number of frames in between each hardware interrupt.
-            // - https://www.alsa-project.org/wiki/FramesPeriods
-            //
-            // FIXME: What values should we set for buffer size and period size? (Should we
-            // set them at all?) virtio-sound spec deals in bytes but ALSA deals
-            // in frames. The alsa bindings sometimes use frames and sometimes bytes.
+            let frame_size = u32::from(s.params.channels)
+                * match s.params.format {
+                    virtio_sound::VIRTIO_SND_PCM_FMT_A_LAW
+                    | virtio_sound::VIRTIO_SND_PCM_FMT_S8
+                    | virtio_sound::VIRTIO_SND_PCM_FMT_U8
+                    | virtio_sound::VIRTIO_SND_PCM_FMT_DSD_U8
+                    | virtio_sound::VIRTIO_SND_PCM_FMT_MU_LAW => 1,
+                    virtio_sound::VIRTIO_SND_PCM_FMT_S16
+                    | virtio_sound::VIRTIO_SND_PCM_FMT_U16
+                    | virtio_sound::VIRTIO_SND_PCM_FMT_DSD_U16
+                    | virtio_sound::VIRTIO_SND_PCM_FMT_IMA_ADPCM => 2,
+                    virtio_sound::VIRTIO_SND_PCM_FMT_S18_3
+                    | virtio_sound::VIRTIO_SND_PCM_FMT_U18_3
+                    | virtio_sound::VIRTIO_SND_PCM_FMT_S20_3
+                    | virtio_sound::VIRTIO_SND_PCM_FMT_U20_3
+                    | virtio_sound::VIRTIO_SND_PCM_FMT_S24_3
+                    | virtio_sound::VIRTIO_SND_PCM_FMT_U24_3
+                    | virtio_sound::VIRTIO_SND_PCM_FMT_S24
+                    | virtio_sound::VIRTIO_SND_PCM_FMT_U24 => 3,
+                    virtio_sound::VIRTIO_SND_PCM_FMT_S20
+                    | virtio_sound::VIRTIO_SND_PCM_FMT_U20
+                    | virtio_sound::VIRTIO_SND_PCM_FMT_S32
+                    | virtio_sound::VIRTIO_SND_PCM_FMT_U32
+                    | virtio_sound::VIRTIO_SND_PCM_FMT_FLOAT
+                    | virtio_sound::VIRTIO_SND_PCM_FMT_DSD_U32
+                    | virtio_sound::VIRTIO_SND_PCM_FMT_IEC958_SUBFRAME => 4,
+                    virtio_sound::VIRTIO_SND_PCM_FMT_FLOAT64 => 8,
+                    _ => unreachable!(),
+                };
+
+            // Calculate desirable bytes/sec rate to achieve the stream's desired
+            // parameters:
+
+            let bps_rate = frame_size * rate;
+
+            // Calculate period size for ~100ms (arbitrary) interrupt period:
+
+            let period_bytes = bps_rate / 10;
+
+            // Finally, calculate the size of a period (in frames):
+
+            let period_frames = period_bytes / frame_size;
+
+            hwp.set_period_size(period_frames as i64, alsa::ValueOr::Less)?;
+
+            // Online ALSA driver recommendations seem to be that the buffer should be at
+            // least 2 * period_size
+            if let Err(err) = hwp.set_buffer_size_near(2 * period_frames as i64) {
+                log::error!("could not set buffer size {}: {}", 2 * period_frames, err);
+            }
+
+            // Read more at https://www.alsa-project.org/wiki/FramesPeriods
 
             pcm.hw_params(&hwp)?;
         }
@@ -133,76 +175,117 @@ fn update_pcm(
     Ok(())
 }
 
-// Returns `true` if the function should be called again, because there are are
-// more data left to write.
+// Returns `Ok(true)` if the function should be called again, because there are
+// are more data left to write.
 fn write_samples_direct(
     pcm: &alsa::PCM,
     stream: &mut Stream,
     mmap: &mut alsa::direct::pcm::MmapPlayback<u8>,
 ) -> AResult<bool> {
     while mmap.avail() > 0 {
-        // Write samples to DMA area from iterator
         let Some(buffer) = stream.buffers.front_mut() else {
             return Ok(false);
         };
-        let mut buf = vec![0; buffer.data_descriptor.len() as usize];
-        let read_bytes = buffer
-            .consume(&mut buf)
-            .expect("failed to read buffer from guest");
-        let mut iter = buf[0..read_bytes as usize].iter().cloned();
+        if !matches!(stream.state, PCMState::Start) {
+            return Ok(false);
+        }
+        if let Err(err) = buffer.prepare_output() {
+            log::error!("Could not read TX buffer, dropping it immediately: {}", err);
+            stream.buffers.pop_front();
+            continue;
+        }
+        // Write samples to DMA area from iterator
+        let mut iter = buffer.bytes[buffer.pos..].iter().cloned();
         let frames = mmap.write(&mut iter);
         let written_bytes = pcm.frames_to_bytes(frames);
         if let Ok(written_bytes) = usize::try_from(written_bytes) {
             buffer.pos += written_bytes;
         }
-        if buffer.pos >= buffer.data_descriptor.len() as usize {
+        if buffer.pos >= buffer.bytes.len() {
             stream.buffers.pop_front();
         }
     }
     match mmap.status().state() {
-        State::Running => {
-            return Ok(false);
-        }
-        State::Prepared => {}
-        State::XRun => {
-            log::trace!("Underrun in audio output stream!");
-            pcm.prepare()?
-        }
-        State::Suspended => {}
+        State::Suspended | State::Running | State::Prepared => Ok(false),
+        State::XRun => Ok(true), // Recover from this in next round
         n => panic!("Unexpected pcm state {:?}", n),
     }
-    Ok(true)
 }
 
+// Returns `Ok(true)` if the function should be called again, because there are
+// are more data left to read.
+fn read_samples_direct(
+    _pcm: &alsa::PCM,
+    stream: &mut Stream,
+    mmap: &mut alsa::direct::pcm::MmapCapture<u8>,
+) -> AResult<bool> {
+    while mmap.avail() > 0 {
+        let Some(buffer) = stream.buffers.front_mut() else {
+            return Ok(false);
+        };
+
+        // Read samples from DMA area with an iterator
+        let mut iter = mmap.iter();
+
+        let mut written_bytes = 0;
+        for (sample, byte) in buffer.bytes[buffer.pos..].iter_mut().zip(&mut iter) {
+            *sample = byte;
+            written_bytes += 1;
+        }
+        buffer.pos += written_bytes as usize;
+        drop(iter);
+        if buffer.pos >= buffer.bytes.len() || mmap.avail() == 0 {
+            stream.buffers.pop_front();
+        }
+    }
+
+    match mmap.status().state() {
+        State::Suspended | State::Running | State::Prepared => Ok(false),
+        State::XRun => Ok(true), // Recover from this in next round
+        n => panic!("Unexpected pcm state {:?}", n),
+    }
+}
+
+// Returns `Ok(true)` if the function should be called again, because there are
+// are more data left to write.
 fn write_samples_io(
     p: &alsa::PCM,
-    stream: &mut Stream,
+    streams: &Arc<RwLock<Vec<Stream>>>,
+    stream_id: usize,
     io: &mut alsa::pcm::IO<u8>,
 ) -> AResult<bool> {
-    loop {
-        let avail = match p.avail_update() {
-            Ok(n) => n,
-            Err(err) => {
-                log::trace!("Recovering from {}", err);
-                p.recover(err.errno() as std::os::raw::c_int, true)?;
-                p.avail_update()?
+    let avail = match p.avail_update() {
+        Ok(n) => n,
+        Err(err) => {
+            log::trace!("Recovering from {}", err);
+            p.recover(err.errno() as std::os::raw::c_int, true)?;
+            if let Err(err) = p.start() {
+                log::error!(
+                    "Could not restart stream {}; ALSA returned: {}",
+                    stream_id,
+                    err
+                );
+                return Err(err);
             }
-        };
-        if avail == 0 {
-            break;
+            p.avail_update()?
         }
-        let written = io.mmap(avail as usize, |buf| {
+    };
+    if avail != 0 {
+        io.mmap(avail as usize, |buf| {
+            let stream = &mut streams.write().unwrap()[stream_id];
             let Some(buffer) = stream.buffers.front_mut() else {
                 return 0;
             };
-            let mut data = vec![0; buffer.data_descriptor.len() as usize];
-
-            // consume() always reads (buffer.data_descriptor.len() -
-            // buffer.pos) bytes
-            let read_bytes = buffer
-                .consume(&mut data)
-                .expect("failed to read buffer from guest");
-            let mut iter = data[0..read_bytes as usize].iter().cloned();
+            if !matches!(stream.state, PCMState::Start) {
+                stream.buffers.pop_front();
+                return 0;
+            }
+            if let Err(err) = buffer.prepare_output() {
+                log::error!("Could not read TX buffer, dropping it immediately: {}", err);
+                stream.buffers.pop_front();
+                return 0;
+            }
+            let mut iter = buffer.bytes[buffer.pos..].iter().cloned();
 
             let mut written_bytes = 0;
             for (sample, byte) in buf.iter_mut().zip(&mut iter) {
@@ -210,21 +293,69 @@ fn write_samples_io(
                 written_bytes += 1;
             }
             buffer.pos += written_bytes as usize;
-            if buffer.pos >= buffer.data_descriptor.len() as usize {
+            if buffer.pos >= buffer.bytes.len() {
                 stream.buffers.pop_front();
             }
             p.bytes_to_frames(written_bytes)
                 .try_into()
                 .unwrap_or_default()
         })?;
-        if written == 0 {
-            break;
-        };
+    } else {
+        return Ok(false);
     }
 
     match p.state() {
-        State::Suspended | State::Running => Ok(false),
-        State::Prepared => Ok(false),
+        State::Suspended | State::Running | State::Prepared => Ok(false),
+        State::XRun => Ok(true), // Recover from this in next round
+        n => panic!("Unexpected pcm state {:?}", n),
+    }
+}
+
+// Returns `Ok(true)` if the function should be called again, because there are
+// are more data left to read.
+fn read_samples_io(
+    p: &alsa::PCM,
+    streams: &Arc<RwLock<Vec<Stream>>>,
+    stream_id: usize,
+    io: &mut alsa::pcm::IO<u8>,
+) -> AResult<bool> {
+    let avail = match p.avail_update() {
+        Ok(n) => n,
+        Err(err) => {
+            log::trace!("Recovering from {}", err);
+            p.recover(err.errno() as std::os::raw::c_int, true)?;
+            if let Err(err) = p.start() {
+                log::error!(
+                    "Could not restart stream {}; ALSA returned: {}",
+                    stream_id,
+                    err
+                );
+                return Err(err);
+            }
+            p.avail_update()?
+        }
+    };
+    if avail == 0 {
+        return Ok(false);
+    }
+    let stream = &mut streams.write().unwrap()[stream_id];
+    let Some(buffer) = stream.buffers.front_mut() else {
+        return Ok(false);
+    };
+    if !matches!(stream.state, PCMState::Start) {
+        stream.buffers.pop_front();
+        return Ok(false);
+    }
+    buffer.prepare_input();
+    let frames_read = io.readi(&mut buffer.bytes[buffer.pos..])?;
+    let bytes_read = p.frames_to_bytes(frames_read as i64);
+    buffer.pos += usize::try_from(bytes_read).unwrap_or_default();
+    if buffer.pos >= buffer.bytes.len() || bytes_read == 0 {
+        stream.buffers.pop_front();
+    }
+
+    match p.state() {
+        State::Suspended | State::Running | State::Prepared => Ok(false),
         State::XRun => Ok(true), // Recover from this in next round
         n => panic!("Unexpected pcm state {:?}", n),
     }
@@ -236,48 +367,68 @@ fn alsa_worker(
     receiver: &Receiver<bool>,
     stream_id: usize,
 ) -> AResult<()> {
+    let direction = streams.write().unwrap()[stream_id].direction;
     loop {
-        let Ok(do_write) = receiver.recv() else {
+        // We get a `true` every time a new I/O message is received from the guest.
+        // If the recv() returns `Ok(false)` or an error, terminate this worker thread.
+        let Ok(do_work) = receiver.recv() else {
             return Ok(());
         };
-        if do_write {
-            loop {
-                if matches!(receiver.try_recv(), Ok(false)) {
-                    break;
-                }
-
+        if do_work {
+            let has_buffers = || -> bool {
+                // Hold `streams` lock as short as possible.
+                let lck = streams.read().unwrap();
+                !lck[stream_id].buffers.is_empty()
+                    && matches!(lck[stream_id].state, PCMState::Start)
+            };
+            // Run this loop till the stream's buffer vector is empty:
+            'empty_buffers: while has_buffers() {
+                // When we return from a read/write attempt and there is still space in the
+                // stream's buffers, get the ALSA file descriptors and poll them till the host
+                // sound device tells us there is more available data.
                 let mut fds = {
                     let lck = pcm.lock().unwrap();
-                    if matches!(lck.state(), State::Running | State::Prepared | State::XRun) {
-                        let mut mmap = lck.direct_mmap_playback::<u8>().ok();
+                    match direction {
+                        Direction::Output => {
+                            let mut mmap = lck.direct_mmap_playback::<u8>().ok();
 
-                        if let Some(ref mut mmap) = mmap {
-                            if write_samples_direct(
-                                &lck,
-                                &mut streams.write().unwrap()[stream_id],
-                                mmap,
-                            )? {
-                                continue;
-                            }
-                        } else {
-                            let mut io = lck.io_bytes();
-                            // Direct mode unavailable, use alsa-lib's mmap emulation instead
-                            if write_samples_io(
-                                &lck,
-                                &mut streams.write().unwrap()[stream_id],
-                                &mut io,
-                            )? {
-                                continue;
+                            if let Some(ref mut mmap) = mmap {
+                                if write_samples_direct(
+                                    &lck,
+                                    &mut streams.write().unwrap()[stream_id],
+                                    mmap,
+                                )? {
+                                    continue 'empty_buffers;
+                                }
+                            } else {
+                                let mut io = lck.io_bytes();
+                                // Direct mode unavailable, use alsa-lib's mmap emulation instead
+                                if write_samples_io(&lck, &streams, stream_id, &mut io)? {
+                                    continue 'empty_buffers;
+                                }
                             }
                         }
-                        lck.get()?
-                    } else {
-                        drop(lck);
-                        sleep(Duration::from_millis(500));
-                        continue;
+                        Direction::Input => {
+                            let mut mmap = lck.direct_mmap_capture::<u8>().ok();
+
+                            if let Some(ref mut mmap) = mmap {
+                                if read_samples_direct(
+                                    &lck,
+                                    &mut streams.write().unwrap()[stream_id],
+                                    mmap,
+                                )? {
+                                    continue 'empty_buffers;
+                                }
+                            } else {
+                                let mut io = lck.io_bytes();
+                                if read_samples_io(&lck, &streams, stream_id, &mut io)? {
+                                    continue 'empty_buffers;
+                                }
+                            }
+                        }
                     }
+                    lck.get()?
                 };
-                // Nothing to do, sleep until woken up by the kernel.
                 alsa::poll::poll(&mut fds, 100)?;
             }
         }
@@ -288,14 +439,15 @@ impl AlsaBackend {
     pub fn new(streams: Arc<RwLock<Vec<Stream>>>) -> Self {
         let (sender, receiver) = channel();
         let sender = Arc::new(Mutex::new(sender));
+        let streams2 = Arc::clone(&streams);
 
         thread::spawn(move || {
-            if let Err(err) = Self::run(streams, receiver) {
+            if let Err(err) = Self::run(streams2, receiver) {
                 log::error!("Main thread exited with error: {}", err);
             }
         });
 
-        Self { sender }
+        Self { sender, streams }
     }
 
     fn run(
@@ -310,7 +462,14 @@ impl AlsaBackend {
             let mut senders = Vec::with_capacity(streams_no);
             for i in 0..streams_no {
                 let (sender, receiver) = channel();
-                let pcm = Arc::new(Mutex::new(PCM::new("default", Direction::Playback, false)?));
+
+                // Initialize with a dummy value, which will be updated every time we call
+                // `update_pcm`.
+                let pcm = Arc::new(Mutex::new(PCM::new(
+                    "default",
+                    Direction::Output.into(),
+                    false,
+                )?));
 
                 let mtx = Arc::clone(&pcm);
                 let streams = Arc::clone(&streams);
@@ -336,11 +495,10 @@ impl AlsaBackend {
 
         while let Ok(action) = receiver.recv() {
             match action {
-                AlsaAction::Read(_) => {}
-                AlsaAction::Write(stream_id) => {
+                AlsaAction::DoWork(stream_id) => {
                     if stream_id >= streams_no {
                         log::error!(
-                            "Received Write action for stream id {} but there are only {} PCM \
+                            "Received DoWork action for stream id {} but there are only {} PCM \
                              streams.",
                             stream_id,
                             pcms.len()
@@ -364,33 +522,20 @@ impl AlsaBackend {
                         );
                         continue;
                     };
-
-                    let start_result = streams.write().unwrap()[stream_id].state.start();
-                    if let Err(err) = start_result {
-                        log::error!("Stream {} start {}", stream_id, err);
-                    } else {
-                        let pcm = &pcms[stream_id];
-                        let lck = pcm.lock().unwrap();
-                        match lck.state() {
-                            State::Running => {}
-                            _ => lck.start()?,
+                    streams.write().unwrap()[stream_id].state.start();
+                    let pcm = &pcms[stream_id];
+                    let lck = pcm.lock().unwrap();
+                    if !matches!(lck.state(), State::Running) {
+                        // Fail gracefully if Start does not succeed.
+                        if let Err(err) = lck.start() {
+                            log::error!(
+                                "Could not start stream {}; ALSA returned: {}",
+                                stream_id,
+                                err
+                            );
                         }
                     }
-                }
-                AlsaAction::Stop(stream_id) => {
-                    if stream_id >= streams_no {
-                        log::error!(
-                            "Received Stop action for stream id {} but there are only {} PCM \
-                             streams.",
-                            stream_id,
-                            pcms.len()
-                        );
-                        continue;
-                    };
-                    let stop_result = streams.write().unwrap()[stream_id].state.stop();
-                    if let Err(err) = stop_result {
-                        log::error!("Stream {} stop {}", stream_id, err);
-                    }
+                    senders[stream_id].send(true).unwrap();
                 }
                 AlsaAction::Prepare(stream_id) => {
                     if stream_id >= streams_no {
@@ -402,15 +547,17 @@ impl AlsaBackend {
                         );
                         continue;
                     };
-                    let prepare_result = streams.write().unwrap()[stream_id].state.prepare();
-                    if let Err(err) = prepare_result {
-                        log::error!("Stream {} prepare {}", stream_id, err);
-                    } else {
-                        let pcm = &pcms[stream_id];
-                        let lck = pcm.lock().unwrap();
-                        match lck.state() {
-                            State::Running => {}
-                            _ => lck.prepare()?,
+                    streams.write().unwrap()[stream_id].state.prepare();
+                    let pcm = &pcms[stream_id];
+                    let lck = pcm.lock().unwrap();
+                    if !matches!(lck.state(), State::Running) {
+                        // Fail gracefully if Prepare does not succeed.
+                        if let Err(err) = lck.prepare() {
+                            log::error!(
+                                "Could not prepare stream {}; ALSA returned: {}",
+                                stream_id,
+                                err
+                            );
                         }
                     }
                 }
@@ -425,15 +572,12 @@ impl AlsaBackend {
                         msg.code = VIRTIO_SND_S_BAD_MSG;
                         continue;
                     };
-                    let release_result = streams.write().unwrap()[stream_id].state.release();
-                    if let Err(err) = release_result {
-                        log::error!("Stream {} release {}", stream_id, err);
-                        msg.code = VIRTIO_SND_S_BAD_MSG;
-                    } else {
-                        senders[stream_id].send(false).unwrap();
-                        let mut streams = streams.write().unwrap();
-                        std::mem::take(&mut streams[stream_id].buffers);
-                    }
+                    // Stop worker thread
+                    senders[stream_id].send(false).unwrap();
+                    // Drop pending stream buffers to complete pending I/O messages
+                    let mut streams_lck = streams.write().unwrap();
+                    streams_lck[stream_id].state.release();
+                    std::mem::take(&mut streams_lck[stream_id].buffers);
                 }
                 AlsaAction::SetParameters(stream_id, mut msg) => {
                     if stream_id >= streams_no {
@@ -456,14 +600,10 @@ impl AlsaBackend {
                     {
                         let mut streams = streams.write().unwrap();
                         let st = &mut streams[stream_id];
-                        if let Err(err) = st.state.set_parameters() {
-                            log::error!("Stream {} set_parameters {}", stream_id, err);
-                            msg.code = VIRTIO_SND_S_BAD_MSG;
-                            continue;
-                        } else if !st.supports_format(request.format)
-                            || !st.supports_rate(request.rate)
-                        {
+                        st.state.set_parameters();
+                        if !st.supports_format(request.format) || !st.supports_rate(request.rate) {
                             msg.code = VIRTIO_SND_S_NOT_SUPP;
+                            // Drop `msg` and skip `update_pcm`
                             continue;
                         } else {
                             st.params.buffer_bytes = request.buffer_bytes;
@@ -473,9 +613,9 @@ impl AlsaBackend {
                             st.params.format = request.format;
                             st.params.rate = request.rate;
                         }
+                        // Manually drop msg for faster response: the kernel has a timeout.
+                        drop(msg);
                     }
-                    // Manually drop msg for faster response: the kernel has a timeout.
-                    drop(msg);
                     update_pcm(&pcms[stream_id], stream_id, &streams)?;
                 }
             }
@@ -514,12 +654,19 @@ macro_rules! send_action {
 
 impl AudioBackend for AlsaBackend {
     send_action! {
-        write Write,
-        read Read,
+        write DoWork,
+        read DoWork,
         prepare Prepare,
         start Start,
-        stop Stop,
     }
+
+    fn stop(&self, id: u32) -> CrateResult<()> {
+        if let Some(s) = self.streams.write().unwrap().get_mut(id as usize) {
+            s.state.stop();
+        }
+        Ok(())
+    }
+
     send_action! {
         ctrl set_parameters SetParameters,
         ctrl release Release,